This is the initial build of the Manos subproject
This is the Manos subproject. It is a simple virtual appliance for Java
test driven development including Ant, JUnit, and Tomcat in a Vagrant
virtual box, with a skeleton application consisting of a servlet, a
class, and a test class. Vagrant Up results (or should) in a working
local build on a 32 bit Centos 6.5 box.